+++ include/route.h
@@ -17,27 +17,27 @@ typedef void (*Handler)(struct epollix_context* ctx);
 
 // Route is a struct that contains the route pattern, handler, and middleware.
 typedef struct Route {
-    HttpMethod method;            // HTTP Method.
-    RouteType type;               // Type of Route (Normal or Static)
-    char* pattern;                // Pattern to match
-    Handler handler;              // Handler for the route
-    PathParams* params;           // Parameters extracted from the URL
-    char* dirname;                // Dirname for static route(dynamic memory)
-    Middleware* middleware;       // Array of middleware functions(allocated dynamically)
-    uint8_t middleware_count;     // Number of middleware functions
-    uint8_t middleware_capacity;  // Capacity of middleware
-    void* mw_data;                // Middleware data. This is set by the user.
+    HttpMethod method;           // HTTP Method.
+    RouteType type;              // Type of Route (Normal or Static)
+    char* pattern;               // Pattern to match
+    Handler handler;             // Handler for the route
+    PathParams* params;          // Parameters extracted from the URL
+    char* dirname;               // Dirname for static route(dynamic memory)
+    Middleware* middleware;      // Array of middleware functions(allocated dynamically)
+    size_t middleware_count;     // Number of middleware functions
+    size_t middleware_capacity;  // Capacity of middleware
+    void* mw_data;               // Middleware data. This is set by the user.
 } Route;
 
 // Route group is a collection of routes that share the same prefix.
 typedef struct RouteGroup {
-    char* prefix;                 // Prefix for the group
-    Route** routes;               // Array of routes(dynamic memory)
-    uint8_t count;                // Number of routes in the group
-    uint8_t capacity;             // capacity of routes in the group
-    Middleware* middleware;       // Middleware for the group
-    uint8_t middleware_count;     // Number of middleware functions
-    uint8_t middleware_capacity;  // Capacity of middleware functions
+    char* prefix;                // Prefix for the group
+    Route** routes;              // Array of routes(dynamic memory)
+    size_t count;                // Number of routes in the group
+    size_t capacity;             // capacity of routes in the group
+    Middleware* middleware;      // Middleware for the group
+    size_t middleware_count;     // Number of middleware functions
+    size_t middleware_capacity;  // Capacity of middleware functions
 } RouteGroup;
 
 // ==================== REGISTER ROUTES ON CTX ===================================
